call hook scripts also before and after processing all interfaces
If ifup or ifdown is called with --all option, before doing anything
to interfaces, if calls all the hook scripts (pre-up or down) with
IFACE set to "--all", LOGICAL set to the current --allow parameter
(or "auto" if it's not set), ADDRFAM="meta" and METHOD="none".

The same is done after all the interfaces have been brought up or taken
down, but different scripts are called, of course.

If implemented properly, existing hook scripts should ignore this change,
as it uses different address family and unusual interface name. More over,
this approach is already used at least by NetworkManager, so any scripts
which could be broken by this change are broken already anyway.

However, to my knowledge there are none scripts which are going to be
broken by this change.